Implement Display and Debug for all public types
